- Made account page a little less annoying to use - Abstracted issue list view controller to separate data from behavior - A little fun caching the projects list with Realm   - Groundwork and experimentation for syncing project users and stuff     to the Realm cache

+++ Facets/Facets/Models/OZLModelIssue.h
@@ -35,9 +35,9 @@
 
 @interface OZLModelIssue : NSObject
 
-@property(nonatomic) int index;
-@property(nonatomic) int projectId;
-@property(nonatomic) int parentIssueId;
+@property(nonatomic) NSInteger index;
+@property(nonatomic) NSInteger projectId;
+@property(nonatomic) NSInteger parentIssueId;
 @property(nonatomic,strong) OZLModelTracker* tracker;
 @property(nonatomic,strong) OZLModelUser* author;
 @property(nonatomic,strong) OZLModelUser* assignedTo;
